merge(5.3-decompose-bid-form): Decompose bid-form-panel (607 lines)
